Fix -Wnarrowing warnings

+++ rebirth/animation.hpp
@@ -15,7 +15,7 @@
 
 
 static inline int
-mix(uint8_t A, uint8_t B, int a, int b, int total) {
+mix(int A, int B, int a, int b, int total) {
     return (A*a + B*b) / (total);
 }
 
@@ -26,9 +26,9 @@ mix(RgbColor a, RgbColor b, int balance) {
     const int bMix = balance;
     const int aMix = total-balance;
     RgbColor ret = {
-        mix(a.r, b.r, aMix, bMix, total),
-        mix(a.g, b.g, aMix, bMix, total),
-        mix(a.b, b.b, aMix, bMix, total),
+        (uint8_t)mix(a.r, b.r, aMix, bMix, total),
+        (uint8_t)mix(a.g, b.g, aMix, bMix, total),
+        (uint8_t)mix(a.b, b.b, aMix, bMix, total),
     };
     return ret;
 }
@@ -101,9 +101,9 @@ scale(uint8_t i, int factor, int max) {
 RgbColor
 scaleBrightness(RgbColor c, int factor, int max) {
     return {
-        scale(c.r, factor, max),
-        scale(c.r, factor, max),
-        scale(c.r, factor, max),
+        (uint8_t)scale(c.r, factor, max),
+        (uint8_t)scale(c.r, factor, max),
+        (uint8_t)scale(c.r, factor, max),
     };
 }
